feat: refactor featured section and add i18n
