+++ frontend/src/app/page.tsx
@@ -1,4 +1,4 @@
-import Featured from '@/components/recommendation/prompt/Featured';
+import Featured from '@/components/recommendation/prompt/featured/Featured';
 import Prompt from '@/components/recommendation/prompt/Prompt';
 import { Container, Box, Typography } from '@mui/material';
 import createTranslation from 'next-translate/useTranslation';
